+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - Log errors that occur in widget event listener blocks to help with troubleshooting
 - disposed? method for Shape objects
 - Add `dispose_images: false, dispose_patterns: false` options to `Shape#dispose` to avoid disposing images/patterns when needed
-
+- Mandelbrot sample background thread pre-calculation of future zoom levels
+- Mandelbrot panning via scrollbars or dragging
 
 ### 4.18.4.8